[util] Fold helper-cairo.cc into helper-cairo.hh
